Move type_intersection functions to TypeUtils.{h,cc}
