+++ opencog/atomutils/Unify.cc
@@ -304,46 +304,6 @@ Handle type_intersection(const Handle& lhs, const Handle& rhs,
 	return Handle::UNDEFINED;
 }
 
-Type type_intersection(Type lhs, Type rhs)
-{
-	ClassServer& cs = classserver();
-	if (cs.isA(lhs, rhs))
-		return lhs;
-	if (cs.isA(rhs, lhs))
-		return rhs;
-	return NOTYPE;              // represent the bottom type
-}
-
-std::set<Type> type_intersection(Type lhs, const std::set<Type>& rhs)
-{
-	std::set<Type> res;
-	// Distribute the intersection over the union type rhs
-	for (Type rhst : rhs) {
-		Type ty = type_intersection(lhs, rhst);
-		if (ty != NOTYPE)
-			res.insert(ty);
-	}
-	return res;
-}
-
-std::set<Type> type_intersection(const std::set<Type>& lhs,
-                                 const std::set<Type>& rhs)
-{
-	// Base cases
-	if (lhs.empty())
-		return rhs;
-	if (rhs.empty())
-		return lhs;
-
-	// Recursive cases
-	std::set<Type> res;
-	for (Type ty : lhs) {
-		std::set<Type> itr = type_intersection(ty, rhs);
-		res.insert(itr.begin(), itr.end());
-	}
-	return res;
-}
-
 std::set<Type> simplify_type_union(std::set<Type>& type)
 {
 	return {}; // TODO: do we really need that?
